Fix synchronization issues between two go routines
We have a go routine that queries repository for a given org and caches
that in redis. Then we have different go routines which query this and
sort the repository list based on different parameters like ForksCount,
IssuesCount and store them into redis again with different key. These
two different go routines need to run one after other and so using a
channel for them to synchronize and communicate in native golang way.